+++ .tryitoutrc.js
@@ -1,0 +1,14 @@
+const { name, description } = require('./package.json');
+
+module.exports = {
+  title: name,
+  description: description,
+  links: {
+    Source: 'https://github.com/gabrielcsapo/npm-what',
+    Download: 'https://github.com/gabrielcsapo/npm-what/releases',
+  },
+  icon: '',
+  demoImage: './example.gif',
+  template: 'product',
+  output: './docs'
+}